refactor comment to issue
