--- github4s/shared/src/main/scala/github4s/free/domain/Comment.scala
@@ -1,28 +1,0 @@
-/*
- * Copyright 2016-2017 47 Degrees, LLC. <http://www.47deg.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package github4s.free.domain
-
-case class Comment(
-    id: Int,
-    url: String,
-    html_url: String,
-    body: String,
-    user: User,
-    created_at: String,
-    updated_at: String)
-
-case class CommentData(body: String)
