+++ github4s/shared/src/test/scala/github4s/utils/BaseSpec.scala
@@ -20,7 +20,6 @@ import cats.Id
 import github4s.GithubResponses.{GHResponse, UnexpectedException}
 import github4s.app.GitHub4s
 import github4s.free.algebra.{
-  CommentOps,
   GitDataOps,
   IssueOps,
   NotificationOps,
@@ -123,6 +122,5 @@ trait BaseSpec extends FlatSpec with Matchers with TestData with IdInstances wit
   class StatusOpsTest       extends StatusOps[GitHub4s]
   class IssueOpsTest        extends IssueOps[GitHub4s]
   class NotificationOpsTest extends NotificationOps[GitHub4s]
-  class CommentOpsTest      extends CommentOps[GitHub4s]
 
 }
